% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R, R/consumeRRS.R
\name{callDTAPI}
\alias{callDTAPI}
\title{HELPER FUNCTION:
This function is a helper that takes in an API key, values in the data table format and column names to pass to the API and the request URL (OData Endpoint Address).
It then obtains a response from Azure Machine Learning Studio and returns a response to the consumeFile function.}
\usage{
callDTAPI(apiKey, requestUrl, columnNames, requestList, globalParam, retryDelay)

callDTAPI(apiKey, requestUrl, columnNames, requestList, globalParam, retryDelay)
}
\description{
HELPER FUNCTION:
This function is a helper that takes in an API key, values in the data table format and column names to pass to the API and the request URL (OData Endpoint Address).
It then obtains a response from Azure Machine Learning Studio and returns a response to the consumeFile function.

This function is a helper that takes in an API key, request URL, column names of the data, request in the data table format (in a lists of lists), global parameters of a web service, and delay time before retrying a call in case of a server error.
It then obtains a response from Azure Machine Learning Studio in the JSON format and returns a response to the consumption functions that call it.
}

