% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R, R/consumeRRS.R
\name{consumeDataframe}
\alias{consumeDataframe}
\title{Consume Data Frame}
\usage{
consumeDataframe(apiKey, requestUrl, scoreDataFrame,
  globalParam = setNames(list(), character(0)), batchSize = 300,
  retryDelay = 0.3)

consumeDataframe(apiKey, requestUrl, scoreDataFrame,
  globalParam = setNames(list(), character(0)), batchSize = 300,
  retryDelay = 0.3)
}
\arguments{
\item{apiKey}{entered as a string}

\item{requestUrl}{entered as a string or discovered through the discover schema method}

\item{scoreDataFrame}{the data frame to be scored}

\item{globalParam}{global parameters entered as a list, default value is an empty list}

\item{batchSize}{batch size of each batch, default value is 250}

\item{retryDelay}{the time in seconds to delay before retrying in case of a server error, default value is 0.3 seconds}

\item{string}{The api key must be entered as the first parameter}

\item{string}{requestURL must be entered as the second parameter}

\item{string}{valuesDF - The name of the data frame that is being scored}

\item{string}{globalParam - global parameters, default value is ""}

\item{int}{batchSize of each batch, which is optional, but 100 by default}

\item{int}{retryDelay the time in seconds to delay before retrying in case of a server error, default value of 0.3 seconds}
}
\value{
results in a list of lists, with the scored probability at the end of each list
# add examples

returnDataFrame data frame containing results returned from web service call
}
\description{
This function takes in an API key, the request URL (OData Endpoint Address), the column names and multiple requests
It scores the experiment with the requests stored in a list of lists, and sends it to the server in the appropriate format.
It then obtains a response from Azure Machine Learning Studio and returns a response to the user. It returns the output column(s) along with the scored probablities!

This function takes in an API key, data frame and the request URL as compulsory parameters.
It scores a data frame and returns results in a data frame.
It calls a helper function that sends requests to the server in the key-value format.
It processes requests in batches and stores the responses in order of batches in an array.
}

