% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R, R/consumeRRS.R
\name{consumeFile}
\alias{consumeFile}
\title{Consume File}
\usage{
consumeFile(apiKey, requestUrl, inFileName, globalParam = setNames(list(),
  character(0)), outputFileName = "results.csv", batchSize = 300,
  retryDelay = 0.3)

consumeFile(apiKey, requestUrl, inFileName, globalParam = setNames(list(),
  character(0)), outputFileName = "results.csv", batchSize = 300,
  retryDelay = 0.3)
}
\arguments{
\item{apiKey}{entered as a string}

\item{requestUrl}{entered as a string or discovered through the discover schema method}

\item{inFileName}{the name of the file scored as a string}

\item{globalParam}{global parameters entered as a list, default value is an empty list}

\item{outputFileName}{the name of the file to output results to, entered as a string, default value is "results.csv"}

\item{batchSize}{batch size of each batch, default value is 250}

\item{retryDelay}{the time in seconds to delay before retrying in case of a server error, default value is 0.3 seconds}

\item{string}{api_key - API key must be entered as the first parameter}

\item{string}{requestURL - must be entered as the third parameter}

\item{string}{infileName - The name of the file that is being scored}

\item{string}{globalParam - global parameters, default value is ""}

\item{string}{outfileName - The name of the file to write results to, has a default value of "results.txt"}

\item{int}{batchSize of each batch, which is optional, but 100 by default}

\item{int}{retryDelay the time in seconds to delay before retrying in case of a server error, default value of 0.3 seconds}
}
\value{
results in a list of lists, with the scored probability at the end of each list
#add example

returnDataFrame data frame containing results returned from web service call
}
\description{
This function takes in an API key, file name and the request URL (OData Endpoint Address).
It calls a helper function that sends requests to the server to the server in the appropriate format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the output columns along with the scored probablities, and stores the result in a text file.

This function takes in an API key, file name and the request URL as compulsory parameters.
It scores a file and returns results in a data frame.
It calls a helper function that sends requests to the server in the key-value format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the results in a data frame, and stores the results in a text file.
}

