% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consumeRRS.R
\name{consumeSingleRows}
\alias{consumeSingleRows}
\title{This function takes in an API key, request URL, column names of the requests and requests as lists in the data table format as compulsory parameters.
It scores single rows of requests and returns results in a data frame.
It calls a helper function that sends requests to the server in the data table format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the results in a data frame.}
\usage{
consumeSingleRows(apiKey, requestUrl, columnNames, ..., globalParam = "",
  retryDelay = 0.3)
}
\arguments{
\item{apiKey}{entered as a string}

\item{requestUrl}{entered as a string or discovered through the discover schema method}

\item{columnNames}{column names entered as a list or discovered through the discover schema method}

\item{...}{variable number of requests entered as rows of values in list format}

\item{globalParam}{global parameters entered as a list, default value is an empty list}

\item{retryDelay}{the time in seconds to delay before retrying in case of a server error, default value is 0.3 seconds}
}
\value{
returnDataFrame data frame containing results returned from web service call
}
\description{
This function takes in an API key, request URL, column names of the requests and requests as lists in the data table format as compulsory parameters.
It scores single rows of requests and returns results in a data frame.
It calls a helper function that sends requests to the server in the data table format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the results in a data frame.
}

