% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consumeRRS.R
\name{discoverSchema}
\alias{discoverSchema}
\title{This function discovers using the workspace ID and web service ID, information specific to the consumption functions}
\usage{
discoverSchema(workspaceId, webserviceID, scheme = "https",
  host = "requestresponse001.cloudapp.net:443", api_version = "2.0")
}
\arguments{
\item{workspaceId}{workspace ID}

\item{webserviceID}{web service ID}

\item{scheme}{the URI scheme}

\item{host}{optional parameter that defaults to ussouthcentral.services.azureml.net}

\item{api_version}{that defaults to 2.0}
}
\value{
List containing the request URL of the webservice, column names of the data, sample input as well as the input schema
}
\description{
This function discovers using the workspace ID and web service ID, information specific to the consumption functions
}

