% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{getDetailsFromUrl}
\alias{getDetailsFromUrl}
\title{This function is a helper that takes in the help URL, and parses the endpoint and workspace from it
This function also documents the assumption that the help URL will end in the format "endpoints/"endpointId/(other keywords)
This function also documents the assumption that the help URL will}
\usage{
getDetailsFromUrl(helpURL)
}
\description{
This function is a helper that takes in the help URL, and parses the endpoint and workspace from it
This function also documents the assumption that the help URL will end in the format "endpoints/"endpointId/(other keywords)
This function also documents the assumption that the help URL will
}

