% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getEndpoints}
\alias{getEndpoints}
\title{Get Endpoints}
\usage{
getEndpoints(wkID, authToken, wsID, url = epURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}
}
\value{
Returns a list of lists, where each endpoint is represented
as a nested named list with the following fields:
"Name", "Description", "CreationTime", "WorkspaceId", "WebServiceId",
"HelpLocation", "PrimaryKey", "SecondaryKey", "ApiLocation", "Version",
"MaxConcurrentCalls", "DiagnosticsTraceLevel", "ThrottleLevel"
}
\description{
Get the endpoints that are part of a web service
}
\examples{
<< Please note that you will need to retrieve all of the signature details from your AzureML account >>
endpoints = getEndpoints("abcdefghijklmnopqrstuvwxyz123456", "abcdefghijklmnopqrstuvwxyz123456", "abcdefghijklmnopqrstuvwxyz123456")
defaultEP = endpoints[[1]]
}

