% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getEPDetails}
\alias{getEPDetails}
\title{Get the details on a specific endpoint}
\usage{
getEPDetails(wkID, authToken, wsID, epID)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}

\item{epID}{The endpoint name}
}
\value{
Returns a named list representing the endpoint with the following fields:
"Name", "Description", "CreationTime", "WorkspaceId", "WebServiceId",
"HelpLocation", "PrimaryKey", "SecondaryKey", "ApiLocation", "Version",
"MaxConcurrentCalls", "DiagnosticsTraceLevel", "ThrottleLevel"
}
\description{
Get the details on a specific endpoint
}
\examples{
defaultEP = getEPDetails("c01fb89129aa4ef0a19affa7f95ecbbc", "523709d06661441bbf129d68f84cd6a4", "6a46d1f2a5e6406b8b1a5c563bf1cd10", "default")
}

