% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getFramework}
\alias{getFramework}
\title{Get FrameWork.}
\usage{
getFramework(tUrl, authToken)
}
\arguments{
\item{tUrl}{The URL from the published web service}

\item{authToken}{The authentication token for the AzureML account being used}
}
\value{
prints the framework
}
\description{
Framework for making an HTTP request to the URL specified returns a list of lists, so that the elements can be accessed via double bracket notation
}
\keyword{Internal}

