% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{getFunctionString}
\alias{getFunctionString}
\title{Get Function Source Code as a String}
\usage{
getFunctionString(x)
}
\arguments{
\item{x}{Name of the function to convert to a string}
}
\value{
function in string format
}
\description{
This is a helper function that will convert a function's source code to a string
Also consider paste(body(fun())) or getAnywhere()
}

