% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getWebServices}
\alias{getWebServices}
\title{Get Web Services}
\usage{
getWebServices(wkID, authToken, url = wsURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}
}
\value{
Returns a list of lists, where each web service is represented
as a nested named list with the following fields:
"Id", "Name", "Description", "CreationTime", "WorkspaceId", "DefaultEndpointName"
}
\description{
Get a list of webservices available to a workspace
}
\examples{
<< Please note that you will need to retrieve all of the signature details from your AzureML account >>
services = getWebServices("abcdefghijklmnopqrstuvwxyz123456", "abcdefghijklmnopqrstuvwxyz123456")
serviceID = services[[1]]["Id"]
}

