% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getWSDetails}
\alias{getWSDetails}
\title{Get Workspace Details.}
\usage{
getWSDetails(wkID, authToken, wsID, url = prodURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}
}
\value{
Returns a list of lists, where each web service is represented as a nested named list with the following fields:

\itemize{
  \item Id
  \item Name
  \item Description
  \item CreationTime
  \item WorkspaceId
  \item DefaultEndpointName
}
}
\description{
Get detailed information about a specific webservice
}
\examples{
\dontrun{
services = getWebServices("abcdefghijklmnopqrstuvwxyz123456", "abcdefghijklmnopqrstuvwxyz123456")
}
}
\seealso{
Other Discovery functions: \code{\link{getEPDetails}};
  \code{\link{getEndpoints}}; \code{\link{getWebServices}}
}

