% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maml-package.R
\docType{package}
\name{maml-package}
\alias{maml}
\alias{maml-package}
\title{Allows you to discover, publish and consume Azure ML web services}
\description{
The discovery code allows the user to retrieve a list of the web services available in their workspace given that they provide the workspace ID and the authorization token (both of which can be found in settings on the AzureML webpage). A user can also get detailed information about a specific web service, retrieve its endpoints, and the details of a specific endpoint.
}
\details{
1. Discovery

\itemize{
  \item Get endpoints: \code{\link{getEndpoints}}
}

2. Publish

blah

3. Consume

blah
}
\keyword{package}

