% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{paramCheck}
\alias{paramCheck}
\title{This is a helper function to check that the user has passed in all of the expected parameters.}
\usage{
paramCheck(userInput, funcName)
}
\arguments{
\item{list}{userInput - List of expected input parameters}

\item{string}{funcName - The function that is being published}
}
\value{
False if the input was not as expected/True if input matched expectation
}
\description{
This is a helper function to check that the user has passed in all of the expected parameters.
}

