% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{predictService}
\alias{predictService}
\title{Make a call to an API with arguments}
\usage{
predictService(serviceUrl = defaultUrl, key = defaultKey,
  toScore = defaultParams)
}
\arguments{
\item{serviceUrl}{The url of the web app}

\item{key}{The API key}

\item{toScore}{the parameters}
}
\value{
The call to the web service at \code{serviceUrl} with arguments \code{toScore}
}
\description{
Make a call to an API with arguments
}
\examples{
nKey = "JlSp5W+RWf2boTHLwvOvW32j/dDI8d/+ghCb8HTZHKYBl+QkZE46w+ZAxTAdo6U1lXfR6G2SBgnK3/i3VznSww=="
nUrl = "https://ussouthcentral.services.azureml.net/workspaces/c01fb89129aa4ef0a19affa7f95ecbbc/services/dadb2d2e626b4e06981dfc6b2b960ebb/execute?api-version=2.0&details=true"
nParams = list(
  Inputs = list(
   "input1" = list(
     "ColumnNames" = list("Column 0", "Class", "Sex", "Age", "Freq"),
     "Values" = list( list( "0", "value", "value", "value", "0" ),  list( "0", "value", "value", "value", "0" )  )
   )                ),
 GlobalParameters = fromJSON('{}')
)
predictService(url, key, params)
}

