% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{publishPreprocess}
\alias{publishPreprocess}
\title{HELPER FUNCTION: Convert Format}
\usage{
publishPreprocess(argList)
}
\arguments{
\item{list}{argList - List of expected input parameters}
}
\value{
Converted inputSchema to the proper format
}
\description{
This is a helper function to convert expected schema to API-expecting format
}

