% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{publishWebServiceOld}
\alias{publishWebServiceOld}
\title{Publish a model to Microsoft Azure}
\usage{
publishWebServiceOld(functionName)
}
\arguments{
\item{functionName}{The function to be published}
}
\value{
The response from the server, including URL and API key
}
\description{
Publish a model to Microsoft Azure
}
\examples{
publishWebService(myFunction
}

