% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{recurDep}
\alias{recurDep}
\title{HELPER FUNCTION: Recursive Dependencies}
\usage{
recurDep(functionName, dependencies, packages)
}
\arguments{
\item{string}{functionName - Name of function to recursively gather dependencies from}

\item{list}{dependencies - List of package dependencies}

\item{list}{packages - Name of available packages}
}
\value{
list of packages and dependencies
}
\description{
This is helper function to recursively gather dependencies from user defined-functions
Similar structure to packDependencies()
recurDep()
}

