% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{recurPkg}
\alias{recurPkg}
\title{HELPER FUNCTION: Recursive Packaging}
\usage{
recurPkg(pkgName, packages)
}
\arguments{
\item{string}{pkgName - Name of package to check for existence in list of packages}

\item{list}{packages - Name of available packages}
}
\value{
list of packages
}
\description{
This is helper function to recursively gather dependencies from user defined-functions
recurPkg()
}

