% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/serialize.R
\name{serializeFunc}
\alias{serializeFunc}
\title{Serialize the Body of a Function}
\usage{
serializeFunc(publishedFunction)
}
\arguments{
\item{string}{function name to serialize}
}
\value{
output (fromJSON)
}
\description{
Change function to accept a dataframe
}

