% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/serialize.R
\name{serializeI}
\alias{serializeI}
\title{Serialize the input}
\usage{
serializeI(input)
}
\arguments{
\item{list}{Take arguments of a function in type list}
}
\value{
serialized input expectations
}
\description{
Create a serialized dataFrame with the input arguments

Expecting:
   inputSchema = list("arg1"="type", "arg2"="type", ...)
}

