% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/serialize.R
\name{serializeMeta}
\alias{serializeMeta}
\title{Serialize (Published) User Arguments}
\usage{
serializeMeta(userDF, funcName)
}
\arguments{
\item{dataframe}{Pass dataframe into function}
}
\value{
serialized output
}
\description{
Take Output from function and save into output dataframe
Expect arguments from user in a dataframe
}

