% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/serialize.R
\name{serializeO}
\alias{serializeO}
\title{Serialize the output}
\usage{
serializeO(output)
}
\arguments{
\item{list}{Take arguments of a function in type list}
}
\value{
serialized output expectations
}
\description{
Serialize the output DataFrame
Expecting:
   outputSchema = list("output1"="type", "output2"="type", ...)
}

