% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucd_checker.R
\name{UCDWords}
\alias{UCDWords}
\title{UCDWords}
\description{
UCDWords

UCDWords
}
\details{
R6 class for managing UCD (Unified Content Descriptor) words.
This class loads the valid UCDs from the `inst/extdata/valid_UCD_v1p6.dat` file
and provides methods to check primary/secondary status, get descriptions,
normalize capitalization, and validate UCD strings.
}
\examples{
ucd <- UCDWords$new()
ucd$is_primary("arith")
ucd$is_secondary("arith.diff")
ucd$get_description("em.ir.15-30um")
ucd$normalize_capitalization("em.ir.15-30um")
ucd$check_ucd("arith;em.IR.8-15um")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UCDWords-new}{\code{UCDWords$new()}}
\item \href{#method-UCDWords-is_primary}{\code{UCDWords$is_primary()}}
\item \href{#method-UCDWords-is_secondary}{\code{UCDWords$is_secondary()}}
\item \href{#method-UCDWords-get_description}{\code{UCDWords$get_description()}}
\item \href{#method-UCDWords-normalize_capitalization}{\code{UCDWords$normalize_capitalization()}}
\item \href{#method-UCDWords-check_ucd}{\code{UCDWords$check_ucd()}}
\item \href{#method-UCDWords-clone}{\code{UCDWords$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UCDWords-new"></a>}}
\if{latex}{\out{\hypertarget{method-UCDWords-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCDWords$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UCDWords-is_primary"></a>}}
\if{latex}{\out{\hypertarget{method-UCDWords-is_primary}{}}}
\subsection{Method \code{is_primary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCDWords$is_primary(ucd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ucd}}{Character. The UCD string to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical. TRUE if the UCD is primary, FALSE otherwise.
Check if a UCD is secondary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UCDWords-is_secondary"></a>}}
\if{latex}{\out{\hypertarget{method-UCDWords-is_secondary}{}}}
\subsection{Method \code{is_secondary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCDWords$is_secondary(ucd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ucd}}{Character. The UCD string to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical. TRUE if the UCD is secondary, FALSE otherwise.
Returns the description of the UCD.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UCDWords-get_description"></a>}}
\if{latex}{\out{\hypertarget{method-UCDWords-get_description}{}}}
\subsection{Method \code{get_description()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCDWords$get_description(ucd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ucd}}{Character. The UCD string to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character.
Returns the standard capitalization of the UCD.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UCDWords-normalize_capitalization"></a>}}
\if{latex}{\out{\hypertarget{method-UCDWords-normalize_capitalization}{}}}
\subsection{Method \code{normalize_capitalization()}}{
Since the standard requires case insensitivity, any capitalization can be passed. This
method returns the standard accepted capitalization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCDWords$normalize_capitalization(ucd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ucd}}{Character. The UCD string to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character.
Check if a UCD string is valid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UCDWords-check_ucd"></a>}}
\if{latex}{\out{\hypertarget{method-UCDWords-check_ucd}{}}}
\subsection{Method \code{check_ucd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCDWords$check_ucd(ucd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ucd}}{Character. The UCD string to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical. TRUE if the UCD is valid, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UCDWords-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UCDWords-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCDWords$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
