% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/serialize.R
\name{unserializeMeta}
\alias{unserializeMeta}
\title{Serialize (Published) Output}
\usage{
unserializeMeta(output)
}
\arguments{
\item{dataframe}{Pass dataframe back to user}
}
\value{
output dataframe
}
\description{
This is an internal function
Take Output from function and save into output dataframe
}

