% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{callAPI}
\alias{callAPI}
\title{This function is a helper that takes in an API key, values in the key value format and column names to pass to the API and the request URL (OData Endpoint Address).
It then obtains a response from Azure Machine Learning Studio and returns a response to the consumeFile function.}
\usage{
callAPI(api_key, requestURL, keyvalues, globalParam, retryDelay)
}
\description{
This function is a helper that takes in an API key, values in the key value format and column names to pass to the API and the request URL (OData Endpoint Address).
It then obtains a response from Azure Machine Learning Studio and returns a response to the consumeFile function.
}

