% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{consumeDataframe}
\alias{consumeDataframe}
\title{This function takes in an API key, the request URL (OData Endpoint Address), the column names and multiple requests
It scores the experiment with the requests stored in a list of lists, and sends it to the server in the appropriate format.
It then obtains a response from Azure Machine Learning Studio and returns a response to the user. It returns the output column(s) along with the scored probablities!}
\usage{
consumeDataframe(api_key, requestURL, valuesDF, globalParam = setNames(list(),
  character(0)), batchSize = 250, retryDelay = 0.3)
}
\arguments{
\item{requestURL}{must be entered as the third parameter, and must be a string}

\item{valuesDF}{the name of the data frame that is being scored}

\item{globalParam}{global parameters entered as a string, default value is ""}

\item{batchSize}{of each batch, which is optional, but 100 by default}

\item{retryDelay}{the time in seconds to delay before retrying in case of a server error, default value of 0.3 seconds}

\item{api}{key must be entered as the first parameter, and must be a string}
}
\value{
results in a list of lists, with the scored probability at the end of each list
}
\description{
This function takes in an API key, the request URL (OData Endpoint Address), the column names and multiple requests
It scores the experiment with the requests stored in a list of lists, and sends it to the server in the appropriate format.
It then obtains a response from Azure Machine Learning Studio and returns a response to the user. It returns the output column(s) along with the scored probablities!
}

