% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{consumeFile}
\alias{consumeFile}
\title{This function takes in an API key, file name and the request URL as compulsory parameters.
It scores a file and returns results in a data frame.
It calls a helper function that sends requests to the server in the key-value format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the results in a data frame, and stores the results in a text file.}
\usage{
consumeFile(apiKey, requestUrl, inFileName, globalParam = setNames(list(),
  character(0)), outputFileName = "results.csv", batchSize = 300,
  retryDelay = 0.3)
}
\arguments{
\item{apiKey}{entered as a string}

\item{requestUrl}{entered as a string or discovered through the discover schema method}

\item{inFileName}{the name of the file scored as a string}

\item{globalParam}{global parameters entered as a list, default value is an empty list}

\item{outputFileName}{the name of the file to output results to, entered as a string, default value is "results.csv"}

\item{batchSize}{batch size of each batch, default value is 250}

\item{retryDelay}{the time in seconds to delay before retrying in case of a server error, default value is 0.3 seconds}
}
\value{
returnDataFrame data frame containing results returned from web service call
}
\description{
This function takes in an API key, file name and the request URL as compulsory parameters.
It scores a file and returns results in a data frame.
It calls a helper function that sends requests to the server in the key-value format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the results in a data frame, and stores the results in a text file.
}

