% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{consumeFile}
\alias{consumeFile}
\title{This function takes in an API key, file name and the request URL (OData Endpoint Address).
It calls a helper function that sends requests to the server to the server in the appropriate format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the output columns along with the scored probablities, and stores the result in a text file.}
\usage{
consumeFile(api_key, requestURL, infileName, globalParam = setNames(list(),
  character(0)), outfileName = "results.csv", batchSize = 250,
  retryDelay = 0.3)
}
\arguments{
\item{requestURL}{must be entered as the third parameter, and must be a string}

\item{infileName}{the name of the file that is being scored}

\item{globalParam}{global parameters entered as a string, default value is ""}

\item{outfileName}{the name of the file to write results to, entered as a string, with a default value of "results.txt"}

\item{batchSize}{of each batch, which is optional, but 100 by default}

\item{retryDelay}{the time in seconds to delay before retrying in case of a server error, default value of 0.3 seconds}

\item{api}{key must be entered as the first parameter, and must be a string}
}
\value{
results in a list of lists, with the scored probability at the end of each list
}
\description{
This function takes in an API key, file name and the request URL (OData Endpoint Address).
It calls a helper function that sends requests to the server to the server in the appropriate format.
It processes requests in batches and stores the responses in order of batches in an array. It returns the output columns along with the scored probablities, and stores the result in a text file.
}

