% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/consume.R
\name{discoverSchema}
\alias{discoverSchema}
\title{This function discovers using the workspace ID and web service ID, information specific to the consumption functions}
\usage{
discoverSchema(helpURL, scheme = "https",
  host = "ussouthcentral.services.azureml.net", api_version = "2.0")
}
\arguments{
\item{helpURL}{help page URL}

\item{scheme}{the URI scheme}

\item{host}{optional parameter that defaults to ussouthcentral.services.azureml.net}

\item{api_version}{that defaults to 2.0}
}
\value{
List containing the request URL of the webservice, column names of the data, sample input as well as the input schema
}
\description{
This function discovers using the workspace ID and web service ID, information specific to the consumption functions
}

