% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getEndpoints}
\alias{getEndpoints}
\title{Get Endpoints.
Get the endpoints that are part of a web service}
\usage{
getEndpoints(wkID, authToken, wsID, url = prodURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}
}
\value{
Returns a list of lists, where each endpoint is represented
as a nested named list with the following fields:
"Name", "Description", "CreationTime", "WorkspaceId", "WebServiceId",
"HelpLocation", "PrimaryKey", "SecondaryKey", "ApiLocation", "Version",
"MaxConcurrentCalls", "DiagnosticsTraceLevel", "ThrottleLevel"
}
\description{
Get Endpoints.
Get the endpoints that are part of a web service
}
\seealso{
For publishing to AzureML, see \code{\link{publishWebService}}

Other Discovery functions: \code{\link{getEPDetails}};
  \code{\link{getWSDetails}}; \code{\link{getWebServices}}
}

