% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getEndpoints}
\alias{getEndpoints}
\title{Get the endpoints that are part of a web service}
\usage{
getEndpoints(wkID, authToken, wsID, url = epURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}
}
\value{
Returns a list of lists, where each endpoint is represented
as a nested named list with the following fields:
"Name", "Description", "CreationTime", "WorkspaceId", "WebServiceId",
"HelpLocation", "PrimaryKey", "SecondaryKey", "ApiLocation", "Version",
"MaxConcurrentCalls", "DiagnosticsTraceLevel", "ThrottleLevel"
}
\description{
Get the endpoints that are part of a web service
}
\examples{
DELETE TOKENS IN REAL VERSION
endpoints = getEndpoints("c01fb89129aa4ef0a19affa7f95ecbbc", "523709d06661441bbf129d68f84cd6a4", "6a46d1f2a5e6406b8b1a5c563bf1cd10")
defaultEP = endpoints[[1]]
}

