% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getEPDetails}
\alias{getEPDetails}
\title{get Endpoint Details}
\usage{
getEPDetails(wkID, authToken, wsID, epName, url = prodURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}

\item{epName}{The endpoint name}
}
\value{
Returns a named list representing the endpoint with the following fields:
"Name", "Description", "CreationTime", "WorkspaceId", "WebServiceId",
"HelpLocation", "PrimaryKey", "SecondaryKey", "ApiLocation", "Version",
"MaxConcurrentCalls", "DiagnosticsTraceLevel", "ThrottleLevel"
}
\description{
Get the details on a specific endpoint
}

