% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getFramework}
\alias{getFramework}
\title{Abstraction for making GET requests.}
\usage{
getFramework(tUrl, authToken)
}
\arguments{
\item{tUrl}{API URL}

\item{authToken}{authentication token}
}
\value{
the response as a named list
}
\description{
Framework for making GET requests to the Azure management APIs.
}
\seealso{
Other discovery: \code{\link{getEPDetails}};
  \code{\link{getEndpoints}}; \code{\link{getWSDetails}};
  \code{\link{getWebServices}}
}
\keyword{Internal}

