% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getWebServices}
\alias{getWebServices}
\title{Get Web Services.}
\usage{
getWebServices(wkID, authToken, url = prodURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}
}
\value{
Returns a list of lists, where each web service is represented as a nested named list with the following fields:

\itemize{
  \item Id
  \item Name
  \item Description
  \item CreationTime
  \item WorkspaceId
  \item DefaultEndpointName
}
}
\description{
Get a list of webservices available to a workspace
}
\seealso{
Other Discovery functions: \code{\link{getEPDetails}};
  \code{\link{getEndpoints}}; \code{\link{getWSDetails}}
}

