% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getWebServices}
\alias{getWebServices}
\title{Get Web Services}
\usage{
getWebServices(wkID, authToken, url = wsURL)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}
}
\value{
Returns a list of lists, where each web service is represented
as a nested named list with the following fields:
"Id", "Name", "Description", "CreationTime", "WorkspaceId", "DefaultEndpointName"
}
\description{
Get a list of webservices available to a workspace
}
\examples{
DELETE TOKENS IN REAL VERSION
services = getWebServices("c01fb89129aa4ef0a19affa7f95ecbbc", "523709d06661441bbf129d68f84cd6a4")
serviceID = services[[1]]["Id"]
}

