% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getWSDetails}
\alias{getWSDetails}
\title{Get detailed information about a specific webservice}
\usage{
getWSDetails(wkID, authToken, wsID)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}
}
\value{
Returns a named list representing the web service
with the following fields:
"Id", "Name", "Description", "CreationTime", "WorkspaceId", "DefaultEndpointName"
}
\description{
Get detailed information about a specific webservice
}
\examples{
service = getWSDetails("c01fb89129aa4ef0a19affa7f95ecbbc", "523709d06661441bbf129d68f84cd6a4", "6a46d1f2a5e6406b8b1a5c563bf1cd10")
serviceID = service["Id"]
}

