% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getWSDetails}
\alias{getWSDetails}
\title{Get Workspace Details}
\usage{
getWSDetails(wkID, authToken, wsID, url = wsURLdet)
}
\arguments{
\item{wkID}{The workspace ID}

\item{authToken}{The primary authorization token}

\item{wsID}{The webservice ID}
}
\value{
Returns a named list representing the web service
with the following fields:
"Id", "Name", "Description", "CreationTime", "WorkspaceId", "DefaultEndpointName"
}
\description{
Get detailed information about a specific webservice
}
\examples{
<< Please note that you will need to retrieve all of the signature details from your AzureML account >>
services = getWebServices("abcdefghijklmnopqrstuvwxyz123456", "abcdefghijklmnopqrstuvwxyz123456")
}

