% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discover.R
\name{getWSDetails}
\alias{getWSDetails}
\title{Get Web Service Details.}
\usage{
getWSDetails(wkID, authToken, wsID, url = prodURL)
}
\arguments{
\item{wkID}{workspace ID}

\item{authToken}{primary authorization token}

\item{wsID}{webservice ID}

\item{url}{the API url to make the call to, by default hits the Azure management API}
}
\value{
Returns a list with the following fields:

\itemize{
  \item Id
  \item Name
  \item Description
  \item CreationTime
  \item WorkspaceId
  \item DefaultEndpointName
}
}
\description{
Get detailed information about a specific Microsoft Azure Machine Learning web service specified by the web service ID.
}
\seealso{
Other discovery: \code{\link{getEPDetails}};
  \code{\link{getEndpoints}}; \code{\link{getFramework}};
  \code{\link{getWebServices}}
}

