% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{packDependencies}
\alias{packDependencies}
\title{This is a helper function to extract object and package dependencies}
\usage{
packDependencies(functionName)
}
\arguments{
\item{string}{functionName - function to package dependencies from}
}
\value{
encoded zip - will return false if nothing was zipped
}
\description{
This is a helper function to extract object and package dependencies
}

