% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publish.R
\name{updateWebService}
\alias{updateWebService}
\title{Update a Published Web Service}
\usage{
updateWebService(functionName, serviceName, wsID, inputSchema, outputSchema,
  wkID, authToken)
}
\arguments{
\item{functionName}{function name as a string to be published}

\item{serviceName}{name of the new web service}

\item{wsID}{ID of the web service to be updated}

\item{inputSchema}{list of input parameters of format \code{list("arg1"="type", "arg2"="type", ...)}}

\item{outputSchema}{list of outputs of format \code{list("output1"="type", "output2"="type", ...)}}

\item{wkID}{the workspace ID}

\item{authToken}{primary authorization token}
}
\value{
List of webservice details, default endpoint details, and the consumption function
}
\description{
This function updates published code given a valid workspace ID and authentication token. The function expects the function name, service id, and the input and output schemas from the user. The user can expect a list of the web service details, the default endpoint details and the consumption function and use this information to access the published function.
}
\seealso{
Other publish: \code{\link{publishWebService}}
}

